helpers removed from Dust core. New npm package created for helpers.
