+++ docs/index.dust.html
@@ -12,7 +12,7 @@
     <script src="lib/dust.js"></script>
     <script src="lib/parser.js"></script>
     <script src="lib/compiler.js"></script>
-    <script src="test/examples.js"></script>
+    <script src="test/jasmine-test/spec/grammarTests.js"></script>
     <script src="test/uutest.js"></script>
     <script src="test/core.js"></script>
     {#tmpl names="select"/}
