+++ dustjs-helpers/test/jasmine-test/client/specRunner.html
@@ -4,18 +4,19 @@
 <head>
   <title>Jasmine Spec Runner</title>
 
-  <link rel="shortcut icon" type="image/png" href="lib/jasmine-1.1.0.rc1/jasmine_favicon.png">
-  <link rel="stylesheet" href="lib/jasmine-1.1.0/jasmine.css" type="text/css" media="screen" title="no title" charset="utf-8">
+  <link rel="shortcut icon" type="image/png" href="../../../../test/jasmine-test/client/lib/jasmine-1.1.0.rc1/jasmine_favicon.png">
+  <link rel="stylesheet" href="../../../../test/jasmine-test/client/lib/jasmine-1.1.0/jasmine.css" type="text/css" media="screen" title="no title" charset="utf-8">
 
-  <script type="text/javascript" src="lib/jasmine-1.1.0/jasmine.js"></script>
-  <script type="text/javascript" src="lib/jasmine-1.1.0/jasmine-html.js"></script>
+  <script type="text/javascript" src="../../../../test/jasmine-test/client/lib/jasmine-1.1.0/jasmine.js"></script>
+  <script type="text/javascript" src="../../../../test/jasmine-test/client/lib/jasmine-1.1.0/jasmine-html.js"></script>
 
   <!-- include spec files here... -->
-  <script src="../spec/grammarTests.js" type="text/javascript" charset="utf-8"></script>
+  <script src="../spec/helpersTests.js" type="text/javascript" charset="utf-8"></script>
   <script src="../spec/renderTestSpec.js" type="text/javascript" charset="utf-8"></script>
 
   <!-- include source files here... -->
-  <script src="../../../dist/dust-full-1.0.0.js" type="text/javascript" charset="utf-8"></script>
+  <script src="../../../../dist/dust-full-1.0.0.js" type="text/javascript" charset="utf-8"></script>
+  <script src="../../../dist/dust-helpers-1.0.0.js" type="text/javascript" charset="utf-8"></script>
 
   <script type="text/javascript">
     (function() {
